backend/requirements.txt
Flask
Flask-CORS
Flask-Caching
python-dotenv
requests
nltk